backend-module/pyproject.toml
[project]
name = "itmo-moodle-agent"
version = "0.1.0"
description = "ITMO Mooddle Agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.13.1",
    "asyncpg>=0.30.0",
    "authlib>=1.6.6",
    "bcrypt==4.3.0",
    "celery>=5.5.3",
    "fastapi>=0.124.4",
    "greenlet>=3.2.4",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.11.0",
    "pydantic[email]>=2.12.3",
    "pyjwt>=2.10.1",
    "pytest-dependency>=0.6.0",
    "python-multipart>=0.0.21",
    "redis>=7.1.0",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-env>=1.2.0",
]
